Add array record dataset helpers with caching
